requirements.txt
numpy
pandas
gurobipy
pydot
automata-lib
scikit-learn
itertools
